requirements.txt
torch
torchvision
tqdm
Pillow
PySide6
PyInstaller
